Adding support for MarkDown and Remark
This commit adds *primitive* support for markdow and remark. With these
packages one can easily modify the pages and prepare presentation from
them. However, the implementation is *REALLY PRIMITIVE*. The code needs
to be extremelly refactored and re-designed. There are some
incompatibilities between the two markdown libraries.

The styling has been adjusted to deal with the incompatibilites in the
`<code>` and `</pre>` parsing. Some pages add directly the `</pre>`
markup. Not ideal, but gets the job done *now*.