Adding of Visit 2.12.0 recipe
Adding the recipe for visit 2.12.0.
In order to install visit, one need to install mesa.
And it depends on the libxml2 with python support. Therefore
the recipe for libxml2 with python support was added.

On daint, visit with easybuild sometimes fails to install when using more
than 16 processors. Not sure why.