Adding xalt to production build
This adds xalt to production build. It will be called only by Jenkins and on daint.

Since we do not have any mechanism to capture if there was a new build we __always__ perform the reverseMap operation.
But we do it in a different folder and then, copy it on top of the production one.