Add QuantumESPRESSO-6.1.0 to daint-gpu software stack
QuantumESPRESSO 5.4.0 is relatively old and shows bugs (e.g.: https://webrt.cscs.ch/Ticket/Display.html?id=27467), therefore I propose to build the more recent release 6.1.0 in the software stack: not yet as default, but that would be the next step to be properly announced to the users in the News section of the User Portal. 
At the same time, I propose to remove the cuda build of release 5.4.0, which is not working.