Updating production build script
This is an attempt to make the production build script more system
agnostic.

./production_builds.sh -h

Usage: ./production_builds.sh [OPTIONS] <list-of-ebfiles>
    -o OS (required)
    -a archtecture (required only on daint)
    -l production list (contains list of eb files)
    -f force the build for a given package
    -h help

The `-o` option should be seen as a prefix folder to define the
installation path. The installation folder is:
$APPS/UES/jenkins/$OS/$ARCH/easybuild

Note that the flag `-a` is only required on daint.

./production_builds.sh -o 6.0.UP02 -a gpu -l CDO -l BOOST

will install all files defines inside the CDO and BOOST files.

./production_builds.sh -o 6.0.UP02 -a gpu -l CDO -l BOOST package1.eb
package2.eb

will add the package1.eb and package2.eb to the build list

./production_builds.sh -o 6.0.UP02 -a gpu -l CDO -l BOOST -f package4.eb
package1.eb package2.eb -f package3.eb

will add the package1.eb and package2.eb to the build list and the
package3.eb and package4.eb will have the additional flag -f added to
their build line.

Note that the -f can be interleaved between other packages and just the
the package after the -f option will be have their build forced.

XALT is only enabled for daint