Removing module purge from daint and dom
I do not understand why we have this module purge. 
It hurts on the builds when attempting to build with craypat support. 
See:
https://github.com/eth-cscs/production/pull/326
https://jenkins.cscs.ch/view/ProductionEB/job/TestingEB/551/label=dom/console

IMHO, there are two solutions, one is to remove the `module purge` and the other is to recover the needed folder by adding `module use ...`. I am choosing the first one if there is no other problem, that I cannot foresee.