Update Visit-2.13.0-CrayGNU-17.08.eb
I had a private discussion with the VisIt developers last night. My testing of the openswr support for rendering was not conclusive but they told me this is normal in this production built which - still - uses an old VTK6 (thus older OpenGL). No much optimization can be obtained. We'll wait for the move to VTK8 to see openswr benefits. Thus, this version build is as good as we can hope.