Update system_wide.cfg
EasyBuild has updated its list of dependencies for installing the `foss` toolchain. This PR hides most of the "unnecessary" ones
```
UCX
GMP
OpenMPI
PMIx
libpciaccess
Perl
DB
groff
makeinfo
libfabric
UnZip
```

This has the potential to affect the non-Cray systems we have.
For some reason that I am not sure, we do not hide GMP on Daint.
So, this PR also affects Daint.